refactor: 중복되는 auth 관련 Type 공통화
