+++ src/types/auth.ts
@@ -1,0 +1,9 @@
+export interface User {
+  email: string;
+  password: string;
+}
+
+export interface UserRegister extends User {
+  nickname: string;
+  birthYear: number;
+}
